Fixes: s/stoi/stol

+++ src/succinct-graph/SuccinctGraph.cpp
@@ -96,7 +96,6 @@ void SuccinctGraph::construct_node_table(const std::string& node_file) {
     this->node_table->serialize();
 }
 
-// FIXME: some stoi should be stol?
 SuccinctGraph& SuccinctGraph::construct(
     std::string node_file,
     std::string edge_file) {
@@ -118,10 +117,10 @@ SuccinctGraph& SuccinctGraph::construct(
         int token_idx = 0;
         while (std::getline(ss, token, ' ')) {
             ++token_idx;
-            if (token_idx == 1) src_id = std::stoi(token);
-            else if (token_idx == 2) dst_id = std::stoi(token);
-            else if (token_idx == 3) atype = std::stoi(token);
-            else if (token_idx == 4) time = std::stoi(token);
+            if (token_idx == 1) src_id = std::stol(token);
+            else if (token_idx == 2) dst_id = std::stol(token);
+            else if (token_idx == 3) atype = std::stol(token);
+            else if (token_idx == 4) time = std::stol(token);
             token.clear();
             if (token_idx == 4) break;
         }
@@ -305,9 +304,9 @@ std::vector<SuccinctGraph::Assoc> SuccinctGraph::assoc_range(
             data_width, curr_off, SuccinctGraphSerde::WIDTH_DATA_WIDTH_PADDED);
         LOG("extracted data width = '%s'\n", data_width.c_str());
 
-        assert(std::stoi(data_width) %
+        assert(std::stol(data_width) %
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width) == 0);
-        cnt = std::stoi(data_width) /
+        cnt = std::stol(data_width) /
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width);
         LOG("cnt = %llu\n", cnt);
 
@@ -396,9 +395,9 @@ std::vector<SuccinctGraph::Assoc> SuccinctGraph::assoc_get(
 
         curr_off += SuccinctGraphSerde::WIDTH_DATA_WIDTH_PADDED;
 
-        assert(std::stoi(data_width) %
+        assert(std::stol(data_width) %
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width) == 0);
-        cnt = std::stoi(data_width) /
+        cnt = std::stol(data_width) /
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width);
         LOG("cnt = %llu\n", cnt);
 
@@ -532,9 +531,9 @@ int64_t SuccinctGraph::assoc_count(int64_t src, int64_t atype) {
         this->edge_table->extract(
             data_width, curr_off, SuccinctGraphSerde::WIDTH_DATA_WIDTH_PADDED);
 
-        assert(std::stoi(data_width) %
+        assert(std::stol(data_width) %
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width) == 0);
-        total_cnt += std::stoi(data_width) /
+        total_cnt += std::stol(data_width) /
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width);
     }
     return total_cnt;
@@ -580,9 +579,9 @@ std::vector<SuccinctGraph::Assoc> SuccinctGraph::assoc_time_range(
 
         curr_off += SuccinctGraphSerde::WIDTH_DATA_WIDTH_PADDED;
 
-        assert(std::stoi(data_width) %
+        assert(std::stol(data_width) %
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width) == 0);
-        cnt = std::stoi(data_width) /
+        cnt = std::stol(data_width) /
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width);
         LOG("cnt = %llu\n", cnt);
 
@@ -747,10 +746,10 @@ void SuccinctGraph::get_neighbors(std::vector<int64_t>& result, int64_t node) {
 
         curr_off += SuccinctGraphSerde::WIDTH_DATA_WIDTH_PADDED;
 
-        assert(std::stoi(data_width) %
+        assert(std::stol(data_width) %
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width) == 0);
 
-        int64_t cnt = std::stoi(data_width) /
+        int64_t cnt = std::stol(data_width) /
             (WIDTH_TIMESTAMP + WIDTH_NODE_ID + edge_width);
 
         curr_off += cnt * WIDTH_TIMESTAMP;
